requirements.txt
numpy==1.19.5
scipy
suds-jurko
casadi==3.5.1
nose==1.3.7
matplotlib
frozendict
PyGeometry-z6
reprep-z6
duckietown-world-daffy==6.2.14
pandas
plotly
kaleido
Pillow